+++ atintegrators/b_drift_kick_drift_expanded.c
@@ -1,2 +1,2 @@
-#include "a_fastdrift.c"
-#include "a_h_k0h_k1h_kn.c"
+#include "drift_fast.h"
+#include "kick_h_k0h_k1h_kn.h"
